实现到删除角色，存在创建角色时token不一致的报错

+++ Unity/Codes/Hotfix/Demo/ServerInfo/ServerInfosComponentSystem.cs
@@ -20,7 +20,7 @@
         public static void Add(this ServerInfosComponent self, ServerInfo serverInfo)
         {
             self.ServerInfoList.Add(serverInfo);
-            self.CurrentServerId = serverInfo.Id;
+            self.CurrentServerId = (int)serverInfo.Id;
         }
     }
 }
